Refactor

+++ CsiRadar.Backend/Core/Entities/CsiData.cs
@@ -17,9 +17,15 @@ public sealed class CsiData
     /// <summary>
     /// Raw interleaved CSI I/Q values directly from the ESP32 payload.
     /// Format: [imaginary_0, real_0, imaginary_1, real_1, ...].
-    /// Amplitude and phase are computed downstream in the Processing layer.
+    /// Each sample is an ESP int8_t (−128..127), so this is <see cref="sbyte"/>[]
+    /// (1 byte/sample). Amplitude and phase are computed downstream in the
+    /// Processing layer; demodulation promotes sbyte→int (max imag²+real² fits int).
+    ///
+    /// NOTE: this array travels through the bounded channel and may be silently
+    /// discarded by <c>BoundedChannelFullMode.DropOldest</c>, so it must NOT be
+    /// rented from an <c>ArrayPool</c> (a dropped frame would never be returned).
     /// </summary>
-    public int[] RawCsiData { get; set; } = [];
+    public sbyte[] RawCsiData { get; set; } = [];
 
     /// <summary>
     /// Number of valid elements in <see cref="RawCsiData"/>.
@@ -30,14 +36,14 @@ public sealed class CsiData
     /// <summary>
     /// Processed CSI amplitude values per subcarrier.
     /// Computed from RawCsiData in the Processing layer via sqrt(I² + Q²).
-    /// Empty until processed by <see cref="Core.Interfaces.ISignalProcessor"/>.
+    /// Empty until processed by <see cref="Core.Interfaces.ICsiStreamProcessor"/>.
     /// </summary>
     public float[] SubcarrierAmplitudes { get; set; } = [];
 
     /// <summary>
     /// Processed CSI phase values (radians) per subcarrier.
     /// Computed from RawCsiData in the Processing layer via atan2(I, Q).
-    /// Empty until processed by <see cref="Core.Interfaces.ISignalProcessor"/>.
+    /// Empty until processed by <see cref="Core.Interfaces.ICsiStreamProcessor"/>.
     /// </summary>
     public float[] SubcarrierPhases { get; set; } = [];
